scripts: fix issue with long interval in function hook_timer
Affected plugins: python, ruby, lua, tcl, guile, javascript, php.

+++ src/plugins/plugin-script-api.c
@@ -523,7 +523,7 @@ plugin_script_api_hook_command_run (struct t_weechat_plugin *weechat_plugin,
 struct t_hook *
 plugin_script_api_hook_timer (struct t_weechat_plugin *weechat_plugin,
                               struct t_plugin_script *script,
-                              int interval, int align_second, int max_calls,
+                              long interval, int align_second, int max_calls,
                               int (*callback)(const void *pointer,
                                               void *data,
                                               int remaining_calls),
